src: Fix warnings in 64-bit build

+++ src/sd_rpc_impl.cpp
@@ -48,7 +48,7 @@ uint32_t sd_rpc_serial_port_enum(sdp_rpc_serial_port_desc_t serial_port_descs[],
         ret = NRF_ERROR_DATA_SIZE; 
     }
 
-	*size = descs.size();
+	*size = (uint32_t) descs.size();
 
     if(ret == NRF_SUCCESS)
     {
